pyproject.toml
[tool.poetry]
name = "wonderland"
version = "0.1.0"
description = ""
authors = ["Henry"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8"
snowflake-snowpark-python = "^0.12.0"
SQLAlchemy = "^1.4.42"
pandas = "^1.5.1"
snowflake-connector-python = "^2.8.0"
tqdm = "^4.64.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
